chore: rename commands to command-maps
